all: clean up code with Go 1.24+ in mind
