+++ src/internal/cm/hostlayout.go
@@ -1,5 +1,3 @@
-//go:build go1.23
-
 package cm
 
 import "structs"
