+++ src/reflect/iter.go
@@ -1,5 +1,3 @@
-//go:build go1.23
-
 // Copyright 2024 The Go Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
